branch complete

+++ src/main/Game.java
@@ -1,13 +1,13 @@
 package main;
 public class Game{
     private String name;
-    private float price;
+    private double price;
     private String supplierID;
     private int uniqueID;
     private double discount;
 
 
-    public Game(String name, float price, String supplierID, int uniqueID, double discount){
+    public Game(String name, double price, String supplierID, int uniqueID, double discount){
         this.name = name;
         this.price = price;
         this.supplierID = supplierID;
@@ -19,7 +19,7 @@ public class Game{
         return this.name;
     }
 
-    public float getPrice(){
+    public double getPrice(){
         return this.price;
     }
 
@@ -35,11 +35,11 @@ public class Game{
 
     // will need to override the .equal() method to check if this is the game for addition of game in inventory
 
-    public float getPriceWithDiscount(boolean saleToggle) { //a more appropriate name would be onSale
-        float chargedPrice = this.price;//if you are using an attribute you don't need to make it a local var
+    public double getPriceWithDiscount(boolean saleToggle) { //a more appropriate name would be onSale
+        double chargedPrice = this.price;//if you are using an attribute you don't need to make it a local var
         if (saleToggle) { //check if auction sale is on and get correct price
             //just make this the return, you don't need thar var
-            chargedPrice = (float) Math.round((price * (1 - this.discount)) * 100) / 100; // decimal places remain at 2
+            chargedPrice = (double) Math.round((price * (1 - this.discount)) * 100) / 100; // decimal places remain at 2
         }
         return chargedPrice;
     }
